Simplify background tasks still further by pulling out a BaseFuture class
