+++ traits_futures/background_call.py
@@ -6,18 +6,14 @@ Background task consisting of a simple callable.
 """
 from traits.api import (
     Any,
-    Bool,
     Callable,
     Dict,
     HasStrictTraits,
-    HasTraits,
-    Instance,
-    on_trait_change,
-    Property,
     Str,
     Tuple,
 )
 
+from traits_futures.base_future import BaseFuture
 from traits_futures.exception_handling import marshal_exception
 from traits_futures.future_states import (
     CANCELLED,
@@ -25,26 +21,10 @@ from traits_futures.future_states import (
     EXECUTING,
     FAILED,
     COMPLETED,
-    WAITING,
-    CANCELLABLE_STATES,
-    DONE_STATES,
-    FutureState,
 )
 from traits_futures.i_job_specification import IJobSpecification
 
-# Message types for messages from CallBackgroundTask to CallFuture.
-# The background task will emit exactly one of the following
-# sequences of message types:
-#
-#   [INTERRUPTED]
-#   [STARTED, RAISED]
-#   [STARTED, RETURNED]
-
-#: Call was cancelled before it started. No arguments.
-INTERRUPTED = "interrupted"
-
-#: Call started executing. No arguments.
-STARTED = "started"
+# The background task sends either a "RAISED" message or a "RETURNED" message.
 
 #: Call failed with an exception. Argument gives exception information.
 RAISED = "raised"
@@ -65,7 +45,6 @@ class CallBackgroundTask:
         self.kwargs = kwargs
 
     def __call__(self, send, cancelled):
-        send(STARTED)
         try:
             result = self.callable(*self.args, **self.kwargs)
         except BaseException as e:
@@ -85,22 +64,10 @@ class CallBackgroundTask:
 # WAITING -> EXECUTING -> COMPLETED
 
 
-class CallFuture(HasStrictTraits):
+class CallFuture(BaseFuture):
     """
     Object representing the front-end handle to a background call.
     """
-
-    #: The state of the background call, to the best of the knowledge of
-    #: this future.
-    state = FutureState
-
-    #: True if this task can be cancelled, else False.
-    cancellable = Property(Bool())
-
-    #: True if we've received the final message from the background task,
-    #: else False. `True` indicates either that the background task
-    #: succeeded, or that it raised, or that it was cancelled.
-    done = Property(Bool())
 
     @property
     def result(self):
@@ -132,26 +99,7 @@ class CallFuture(HasStrictTraits):
             raise AttributeError("No exception has been raised for this call.")
         return self._exception
 
-    def cancel(self):
-        """
-        Method that can be called from the main thread to
-        indicate that the task should be cancelled (provided
-        it hasn't already started running).
-        """
-        # In the interests of catching coding errors early in client
-        # code, we're strict about what states we allow cancellation
-        # from. Some applications may want to weaken the error below
-        # to a warning, or just do nothing on an invalid cancellation.
-        if not self.cancellable:
-            raise RuntimeError("Can only cancel a queued or executing task.")
-        self._cancel()
-        self.state = CANCELLING
-
     # Private traits ##########################################################
-
-    #: Callable called with no arguments to request cancellation of the
-    #: background task.
-    _cancel = Callable()
 
     #: Result from the background task.
     _result = Any()
@@ -159,25 +107,7 @@ class CallFuture(HasStrictTraits):
     #: Exception information from the background task.
     _exception = Tuple(Str(), Str(), Str())
 
-    #: Object that receives messages from the background task.
-    _message_receiver = Instance(HasTraits)
-
     # Private methods #########################################################
-
-    @on_trait_change("_message_receiver:message")
-    def _process_message(self, message):
-        message_type, message_arg = message
-        method_name = "_process_{}".format(message_type)
-        getattr(self, method_name)(message_arg)
-
-    def _process_interrupted(self, none):
-        assert self.state in (CANCELLING,)
-        self.state = CANCELLED
-
-    def _process_started(self, none):
-        assert self.state in (WAITING, CANCELLING)
-        if self.state == WAITING:
-            self.state = EXECUTING
 
     def _process_raised(self, exception_info):
         assert self.state in (EXECUTING, CANCELLING)
@@ -194,25 +124,6 @@ class CallFuture(HasStrictTraits):
             self.state = COMPLETED
         else:
             self.state = CANCELLED
-
-    def _get_cancellable(self):
-        return self.state in CANCELLABLE_STATES
-
-    def _get_done(self):
-        return self.state in DONE_STATES
-
-    def _state_changed(self, old_state, new_state):
-        old_cancellable = old_state in CANCELLABLE_STATES
-        new_cancellable = new_state in CANCELLABLE_STATES
-        if old_cancellable != new_cancellable:
-            self.trait_property_changed(
-                "cancellable", old_cancellable, new_cancellable
-            )
-
-        old_done = old_state in DONE_STATES
-        new_done = new_state in DONE_STATES
-        if old_done != new_done:
-            self.trait_property_changed("done", old_done, new_done)
 
 
 @IJobSpecification.register
