+++ traits_futures/background_iteration.py
@@ -8,19 +8,15 @@ import types
 
 from traits.api import (
     Any,
-    Bool,
     Callable,
     Dict,
     Event,
     HasStrictTraits,
-    HasTraits,
-    Instance,
-    on_trait_change,
-    Property,
     Str,
     Tuple,
 )
 
+from traits_futures.base_future import BaseFuture
 from traits_futures.exception_handling import marshal_exception
 from traits_futures.future_states import (
     CANCELLED,
@@ -29,9 +25,6 @@ from traits_futures.future_states import (
     FAILED,
     COMPLETED,
     WAITING,
-    CANCELLABLE_STATES,
-    DONE_STATES,
-    FutureState,
 )
 from traits_futures.i_job_specification import IJobSpecification
 
@@ -50,9 +43,6 @@ from traits_futures.i_job_specification import IJobSpecification
 #: Iteration was cancelled either before it started or during the
 #: iteration. No arguments.
 INTERRUPTED = "interrupted"
-
-#: Iteration started executing. No arguments.
-STARTED = "started"
 
 #: Iteration failed with an exception, or there was
 #: an exception on creation of the iterator. Argument gives
@@ -78,7 +68,6 @@ class IterationBackgroundTask:
         self.kwargs = kwargs
 
     def __call__(self, send, cancelled):
-        send(STARTED)
         try:
             iterable = iter(self.callable(*self.args, **self.kwargs))
         except BaseException as e:
@@ -131,23 +120,11 @@ class IterationBackgroundTask:
 # no results events will be fired after cancelling.
 
 
-class IterationFuture(HasStrictTraits):
+class IterationFuture(BaseFuture):
     """
     Foreground representation of an iteration executing in the
     background.
     """
-
-    #: The state of the background iteration, to the best of the knowledge of
-    #: this future.
-    state = FutureState
-
-    #: True if this task can be cancelled, else False.
-    cancellable = Property(Bool())
-
-    #: True if we've received the final message from the background iteration,
-    #: else False. `True` indicates either that the background iteration
-    #: succeeded, or that it raised, or that it was cancelled.
-    done = Property(Bool())
 
     #: Event fired whenever a result arrives from the background
     #: iteration.
@@ -168,48 +145,12 @@ class IterationFuture(HasStrictTraits):
             raise AttributeError("No exception has been raised for this call.")
         return self._exception
 
-    def cancel(self):
-        """
-        Method called from the main thread to request cancellation
-        of the background job.
-        """
-        # In the interests of catching coding errors early in client
-        # code, we're strict about what states we allow cancellation
-        # from. Some applications may want to weaken the error below
-        # to a warning, or just do nothing on an invalid cancellation.
-        if not self.cancellable:
-            raise RuntimeError("Can only cancel a queued or executing task.")
-        self._cancel()
-        self.state = CANCELLING
-
     # Private traits ##########################################################
-
-    #: Callable called with no arguments to request cancellation of the
-    #: background task.
-    _cancel = Callable()
 
     #: Exception information from the background task.
     _exception = Tuple(Str(), Str(), Str())
 
-    #: Object that receives messages from the background task.
-    _message_receiver = Instance(HasTraits)
-
     # Private methods #########################################################
-
-    @on_trait_change("_message_receiver:message")
-    def _process_message(self, message):
-        message_type, message_arg = message
-        method_name = "_process_{}".format(message_type)
-        getattr(self, method_name)(message_arg)
-
-    def _process_interrupted(self, none):
-        assert self.state in (CANCELLING,)
-        self.state = CANCELLED
-
-    def _process_started(self, none):
-        assert self.state in (WAITING, CANCELLING)
-        if self.state == WAITING:
-            self.state = EXECUTING
 
     def _process_raised(self, exception_info):
         assert self.state in (WAITING, EXECUTING, CANCELLING)
@@ -232,25 +173,6 @@ class IterationFuture(HasStrictTraits):
         # Any results arriving after a cancellation request are ignored.
         if self.state == EXECUTING:
             self.result_event = result
-
-    def _get_cancellable(self):
-        return self.state in CANCELLABLE_STATES
-
-    def _get_done(self):
-        return self.state in DONE_STATES
-
-    def _state_changed(self, old_state, new_state):
-        old_cancellable = old_state in CANCELLABLE_STATES
-        new_cancellable = new_state in CANCELLABLE_STATES
-        if old_cancellable != new_cancellable:
-            self.trait_property_changed(
-                "cancellable", old_cancellable, new_cancellable
-            )
-
-        old_done = old_state in DONE_STATES
-        new_done = new_state in DONE_STATES
-        if old_done != new_done:
-            self.trait_property_changed("done", old_done, new_done)
 
 
 @IJobSpecification.register
